Fix HA tests
