+++ tests/integration_test/src/utils.py
@@ -19,6 +19,7 @@ import shutil
 import subprocess
 import sys
 import tempfile
+import time
 from typing import List
 
 import yaml
@@ -28,6 +29,7 @@ from nvflare.fuel.hci.client.api_status import APIStatus
 from nvflare.fuel.hci.client.fl_admin_api import FLAdminAPI
 from nvflare.fuel.hci.client.fl_admin_api_constants import FLDetailKey
 from nvflare.fuel.hci.client.fl_admin_api_spec import TargetType
+from nvflare.fuel.utils.class_utils import instantiate_class
 
 from .constants import DEFAULT_RESOURCE_CONFIG, FILE_STORAGE, PROVISION_SCRIPT, RESOURCE_CONFIG
 from .example import Example
@@ -377,3 +379,61 @@ def generate_test_config_yaml_for_example(
             yaml.dump(config, yaml_file, default_flow_style=False)
         output_yamls.append(output_yaml)
     return output_yamls
+
+
+def _read_admin_json_file(admin_json_file) -> dict:
+    if not os.path.exists(admin_json_file):
+        raise RuntimeError("Missing admin json file.")
+    with open(admin_json_file, "r") as f:
+        admin_json = json.load(f)
+    return admin_json
+
+
+def create_admin_api(workspace_root_dir, upload_root_dir, download_root_dir, admin_user_name, poc):
+    admin_startup_folder = os.path.join(workspace_root_dir, admin_user_name, "startup")
+    admin_json_file = os.path.join(admin_startup_folder, "fed_admin.json")
+    admin_json = _read_admin_json_file(admin_json_file)
+    overseer_agent = instantiate_class(
+        class_path=admin_json["admin"]["overseer_agent"]["path"],
+        init_params=admin_json["admin"]["overseer_agent"]["args"],
+    )
+
+    ca_cert = ""
+    client_key = ""
+    client_cert = ""
+    if not poc:
+        ca_cert = os.path.join(admin_startup_folder, admin_json["admin"]["ca_cert"])
+        client_key = os.path.join(admin_startup_folder, admin_json["admin"]["client_key"])
+        client_cert = os.path.join(admin_startup_folder, admin_json["admin"]["client_cert"])
+
+    admin_api = FLAdminAPI(
+        upload_dir=upload_root_dir,
+        download_dir=download_root_dir,
+        overseer_agent=overseer_agent,
+        poc=poc,
+        user_name=admin_user_name,
+        ca_cert=ca_cert,
+        client_key=client_key,
+        client_cert=client_cert,
+        auto_login_max_tries=20,
+    )
+    return admin_api
+
+
+def ensure_admin_api_logged_in(admin_api: FLAdminAPI, timeout: int = 60):
+    login_success = False
+    try:
+        start_time = time.time()
+        while time.time() - start_time <= timeout:
+            if admin_api.is_ready():
+                login_success = True
+                break
+            time.sleep(0.2)
+
+        if not login_success:
+            print(f"Admin api failed to log in within {timeout} seconds: {admin_api.fsm.current_state}.")
+        else:
+            print("Admin successfully logged into server.")
+    except Exception as e:
+        print(f"Exception in logging in to admin: {e.__str__()}")
+    return login_success
